aws-gcp-wif-connection/src/lambdas/big_query_test/requirements.txt
google-auth>=2.35.0
google-cloud-bigquery>=3.40.0
aws-lambda-powertools>=3.24.0
